Changed variable (max_count_elements) in the main struct

+++ srcs/init.c
@@ -7,6 +7,6 @@ void	ft_init(t_vector	*vector, unsigned short size_element)
 	vector->storage = NULL;
 	vector->size_element = size_element;
 	vector->count_elements = 0;
-	vector->max_count_elements = SIZE_MAX;
+	vector->max_count_elements = INT32_MAX;
 	vector->capacity = 0;
 }
